+++ site/contentlayer.config.ts
@@ -5,7 +5,7 @@ import { rehypeMetaAttribute } from './lib/rehype-meta-attribute';
 
 export const Doc = defineDocumentType(() => ({
   name: 'Doc',
-  filePathPattern: 'docs/**/*.mdx',
+  filePathPattern: '**/docs/**/*.mdx',
   contentType: 'mdx',
   fields: {
     title: {
@@ -20,14 +20,18 @@ export const Doc = defineDocumentType(() => ({
   computedFields: {
     slug: {
       type: 'string',
-      resolve: (doc) => doc._raw.flattenedPath.replace('docs/', ''),
+      resolve: (doc) => doc._raw.flattenedPath.split('/').slice(2).join('/'),
+    },
+    locale: {
+      type: 'string',
+      resolve: (doc) => doc._raw.sourceFilePath.split('/')[0],
     },
   },
 }));
 
 export const Guide = defineDocumentType(() => ({
   name: 'Guide',
-  filePathPattern: 'guides/**/*.mdx',
+  filePathPattern: '**/guides/**/*.mdx',
   contentType: 'mdx',
   fields: {
     title: {
@@ -46,7 +50,12 @@ export const Guide = defineDocumentType(() => ({
   computedFields: {
     slug: {
       type: 'string',
-      resolve: (doc) => doc._raw.flattenedPath.replace('guides/', ''),
+      resolve: (guide) =>
+        guide._raw.flattenedPath.split('/').slice(2).join('/'),
+    },
+    locale: {
+      type: 'string',
+      resolve: (guide) => guide._raw.sourceFilePath.split('/')[0],
     },
   },
 }));