+++ site/next.config.js
@@ -8,6 +8,25 @@ const withVanillaExtract = createVanillaExtractPlugin();
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
+  i18n: {
+    locales: [
+      'en-US',
+      'ar',
+      'es-419',
+      'fr',
+      'hi',
+      'id',
+      'ja',
+      'ko',
+      'pt-BR',
+      'ru',
+      'th',
+      'tr',
+      'zh-CN',
+    ],
+    defaultLocale: 'en-US',
+    localeDetection: true,
+  },
   async redirects() {
     return [
       {