+++ packages/create-rainbowkit/generated-test-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # generated-test-app
 
+## 0.1.73
+
+### Patch Changes
+
+- Updated dependencies [ef64a229]
+  - @rainbow-me/rainbowkit@1.2.0
+
+## 0.1.72
+
+### Patch Changes
+
+- Updated dependencies [9f68c300]
+- Updated dependencies [3f595c12]
+- Updated dependencies [e2075b31]
+  - @rainbow-me/rainbowkit@1.1.4
+
 ## 0.1.71
 
 ### Patch Changes