+++ packages/example/CHANGELOG.md
@@ -1,5 +1,23 @@
 # example
 
+## 0.0.80
+
+### Patch Changes
+
+- Updated dependencies [ef64a229]
+  - @rainbow-me/rainbowkit@1.2.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.3.2
+
+## 0.0.79
+
+### Patch Changes
+
+- Updated dependencies [9f68c300]
+- Updated dependencies [3f595c12]
+- Updated dependencies [e2075b31]
+  - @rainbow-me/rainbowkit@1.1.4
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.3.1
+
 ## 0.0.78
 
 ### Patch Changes
@@ -38,7 +56,7 @@
 
 - Updated dependencies [b37f5d68]
   - @rainbow-me/rainbowkit@1.1.0
-  - @rainbow-me/rainbowkit-siwe-next-auth@1.0.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.3.2
 
 ## 0.0.74